pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bandcon"
version = "0.0.1"
description = "Band-Supervised Contrastive Learning for Inverse RNA Circuit Design"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Anonymous Authors"}]
dependencies = [
  "numpy>=1.22",
  "scipy>=1.8",
  "torch>=2.1; platform_system != 'Darwin' or platform_machine != 'arm64'",
  "torch>=2.1; platform_system == 'Darwin' and platform_machine == 'arm64'",
  "pytorch-lightning>=2.1",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "tqdm>=4.64",
  "pyyaml>=6.0",
  "networkx>=3.0",
  "pandas>=1.5",
  "matplotlib>=3.6",
  "jax>=0.4.16",
  "jaxlib>=0.4.16",
  "ipykernel>=6.30.1"
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.4.0", "mypy>=1.8"]

[tool.setuptools.packages.find]
where = ["."]
include = ["bandcon*"]
exclude = ["tests*"]
